moving ObjTextList in it's own cpp/hpp files
